requirements.txt
eventlet==0.32.0
python-socketio==5.4.0
aiohttp==3.7.4.post0
